Threadsafe PathStore; Compiles with strict compiler flags
path_store_thread_safe.c:
* Sketched out kmer level locking for adding paths => removes global path_store lock

libs/Makefile: fixed build broken by renaming hash_functions -> misc

Compiled with strict compiler flags and fixed all issues
* lot of type fixed

util.c:parse_entire_int() family of functions now take const arguments

hash_table.c: fixed prefetch so data doesn't linger in the cache

+++ src/basic/file_filter.c
@@ -50,7 +50,7 @@ boolean file_filter_alloc(FileFilter *fltr, char *path,
   strbuf_set(&fltr->orig_path, path);
 
   file_filter_deconstruct_path(path, &path_start, &path_end);
-  fltr->intocol = (path_start == path ? 0 : atoi(path));
+  fltr->intocol = (size_t)(path_start == path ? 0 : atoi(path));
 
   path_lchar = *path_end;
   *path_end = '\0';
